Make subs and words controllers query the datbase directly
This took quite a while because as it turns out, mongodb (much like
protocol buffers) can't really handle arrays of arrays.

Changing the schema to an array of objects that contain arrays
fixed the problem.

+++ backend/Erabikata.Models/Output/WordInfo.cs
@@ -13,7 +13,7 @@ namespace Erabikata.Models.Output
         [JsonProperty(Required = Required.Always)] public int Rank { get; set; } = -1;
 
         [JsonProperty(Required = Required.Always)]
-        public int TotalOccurrences { get; set; }
+        public long TotalOccurrences { get; set; }
 
         [JsonProperty(Required = Required.Always)]
         public IEnumerable<Occurence> Occurrences { get; set; } = Enumerable.Empty<Occurence>();
